LPD-65879 Refactor. Migrate viewsReducer to typescript
